+++ src/main/java/roomescape/payment/service/PaymentService.java
@@ -5,7 +5,6 @@ import roomescape.payment.dto.PaymentResponse;
 
 public interface PaymentService {
 
-    PaymentResponse createOrder(PaymentRequest paymentRequest);
 
     PaymentResponse confirmPayment(PaymentRequest paymentRequest);
